+++ bin/install_uninstall.ml
@@ -18,13 +18,12 @@ let get_dirs context ~prefix_from_command_line ~libdir_from_command_line =
     Fiber.return (prefix, Some (Path.relative prefix dir))
   | None ->
     let open Fiber.O in
-    let* prefix = Memo.Build.run (Context.install_prefix context) in
-    let libdir =
+    let* prefix = Context.install_prefix context in
+    let+ libdir =
       match libdir_from_command_line with
       | None -> Memo.Build.run (Context.install_ocaml_libdir context)
       | Some l -> Fiber.return (Some (Path.relative prefix l))
     in
-    let+ libdir = libdir in
     (prefix, libdir)
 
 let resolve_package_install setup pkg =