+++ otherlibs/dune-rpc/private/types.ml
@@ -32,7 +32,7 @@ end
 module Version = struct
   type t = int * int
 
-  let latest = (1, 0)
+  let latest = (3, 0)
 
   let sexp : t Conv.value =
     let open Conv in