+++ src/dune_engine/stanza.ml
@@ -6,7 +6,12 @@ module Parser = struct
   type nonrec t = string * t list Dune_lang.Decoder.t
 end
 
-let latest_version = (3, 0)
+(* The actual latest version is defined in the rpc library. This is because rpc
+   client needs to know the version of dune to use to connect.
+
+   To upgrade the latest version of the dune language, you need to edit the file
+   in the rpc library. *)
+let latest_version = Dune_rpc_private.Version.latest
 
 let since v = (v, `Since v)