+++ src/dune_rules/findlib/meta.ml
@@ -163,6 +163,11 @@ let builtins ~stdlib_dir ~version:ocaml_version =
       | Some a -> a
     in
     let main_modules =
+      let name =
+        String.map name ~f:(function
+          | '-' -> '_'
+          | c -> c)
+      in
       if labels then
         main_modules [ name; name ^ "Labels" ]
       else
@@ -203,7 +208,7 @@ let builtins ~stdlib_dir ~version:ocaml_version =
         ; sub "bytecomp" [ "compiler-libs.common" ]
         ; sub "optcomp" [ "compiler-libs.common" ]
         ; sub "toplevel" [ "compiler-libs.bytecomp" ] ~kind:[ Byte ]
-        ; sub "toplevel"
+        ; sub "native-toplevel"
             [ "compiler-libs.optcomp"; "dynlink" ]
             ~kind:[ Native ]
             ~exists_if_ext:(Mode.compiled_lib_ext Native)